- Marging and Refactoring to cover issue 20 and 21
Necessary refactoring to achieve the goal related to issue 20 and 21, based on a same infrastructure. - The report json file is already been generating as well as the JMX object named as org.lareferencia:type=EntityLoadingMonitorService. Everything based on the dry-run mode. The validation handler over the non dry-run mode still pendent to be implemented.